# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Rest RCS Express",
    "description": "API Rest RCS Express",
    "contact": {
      "name": "rcsxpress",
      "email": "operaciones@rcsexpress.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Ambiente de Desarrollo"
    },
    {
      "url": "https://scen.rcsexpress.com/api/v1/",
      "description": "Ambiente de Producción"
    }
  ],
  "consumes": ["application/json"],
  "produces": ["application/json"],
...
  "paths": {
...
    "/paises": {
      "get": {
        "tags": ["Paises"],
        "summary": "Obtener todos los paises",
        "description": "API para Obtener todos los paises del sistema",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paises"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Paises"],
        "summary": "Crear Paises",
        "description": "API para Crear un nuevo Pais",
        "requestBody": {
          "description": "Post object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Paises"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paises"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "415": {
            "description": "MediaType No Soportado"
          },
          "500": {
            "description": "Error Interno del Sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/500"
                }
              }
            }
          }
        }
      }
    },
    "/paises/{id}": {
      "get": {
        "tags": ["Paises"],
        "summary": "Obtener Pais por su Id",
        "description": "API para Retornar un Pais por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo del Pais",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paises"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Paises"],
        "summary": "Actualizar Pais",
        "description": "API para Actualizar un Pais por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo del Pais",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Put object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Paises"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paises"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "415": {
            "description": "MediaType No Soportado"
          },
          "500": {
            "description": "Error Interno del Sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/500"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Paises"],
        "summary": "Borrar Pais",
        "description": "API para Borrar un Pais por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo del Pais",
            "required": true, "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "$ref": "#/components/schemas/Paises/properties/id"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      }
    },
...
    "/permisos": {
      "get": {
        "tags": ["Permisos"],
        "summary": "Obtener todos los permisos",
        "description": "API para Obtener todos los permisos del sistema",
        "parameters": [
          {
            "name": "rol",
            "in": "header",
            "description": "Codigo del Rol",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permisos"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Permisos"],
        "summary": "Crear Permiso",
        "description": "API para Crear un nuevo Permiso",
        "requestBody": {
          "description": "Post object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Permisos"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permisos"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "415": {
            "description": "MediaType No Soportado"
          },
          "500": {
            "description": "Error Interno del Sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/500"
                }
              }
            }
          }
        }
      }
    },
    "/permisos/{id}": {
      "get": {
        "tags": ["Permisos"],
        "summary": "Obtener Permiso por su Id",
        "description": "API para Retornar un Permiso por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo del Permiso",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permisos"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Permisos"],
        "summary": "Actualizar Permiso",
        "description": "API para Actualizar un Permiso por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo del Permiso",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Put object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Permisos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permisos"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "415": {
            "description": "MediaType No Soportado"
          },
          "500": {
            "description": "Error Interno del Sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/500"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Permisos"],
        "summary": "Borrar Permiso",
        "description": "API para Borrar un Permiso por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo del Permiso",
            "required": true, "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "$ref": "#/components/schemas/Permisos/properties/id"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Paises": {
        "description": "Modelo de Paises",
        "required": ["id", "desc_pais", "tipo_pais"],
        "properties": {
          "id": {
            "type": "number",
            "readOnly": true
          },
          "desc_pais": {
            "type": "string"
          },
          "tipo_pais": {
            "type": "string"
          },
          "tipo_pais_desc": {
            "type": "string",
            "readOnly": true
          }
        }
      },
...
      "Permisos": {
        "description": "Modelo de Permisos",
        "required": ["id", "codigo", "cod_rol"],
        "properties": {
          "id": {
            "type": "number",
            "readOnly": true
          },
          "codigo": {
            "type": "string"
          },
          "cod_rol": {
            "type": "number"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}